# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MS Elevate LEAPS Tracker API
  version: 1.0.0
  description: |
    Complete API documentation for the Microsoft Elevate Indonesia LEAPS Tracker platform.
    
    The LEAPS framework consists of 5 stages:
    - **Learn**: Complete Microsoft Learn courses (20 points)
    - **Explore**: Apply AI in classroom with evidence (50 points)  
    - **Amplify**: Train peers/students (2 points/peer, 1 point/student)
    - **Present**: LinkedIn post with screenshot (20 points)
    - **Shine**: Innovation ideas for recognition
    
    ## Authentication
    All authenticated endpoints require a valid Clerk JWT token in the Authorization header.
    
    ## Rate Limits
    - File uploads: 10MB max per file
    - Amplify submissions: 50 peers, 200 students per 7-day rolling window
    - Bulk operations: Maximum 50 items per request
    
  contact:
    name: MS Elevate Team
    url: https://leaps.mereka.org
servers:
  - url: https://leaps.mereka.org/api
    description: Production server
  - url: https://staging-leaps.mereka.org/api 
    description: Staging server
security:
  - ClerkJWT: []
components:
  securitySchemes:
    ClerkJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk authentication JWT token
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        details:
          type: array
          items:
            type: object
          description: Validation error details (for Zod errors)
      required:
        - error
    User:
      type: object
      properties:
        id:
          type: string
        handle:
          type: string
        name:
          type: string
        email:
          type: string
        avatar_url:
          type: string
        school:
          type: string
        cohort:
          type: string
        role:
          type: string
          enum: [participant, reviewer, admin, superadmin]
        profile_visibility:
          type: string
          enum: [PUBLIC, PRIVATE]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - handle
        - name
        - email
...
    Submission:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
        activity_code:
          type: string
          enum: [LEARN, EXPLORE, AMPLIFY, PRESENT, SHINE]
        status:
          type: string
          enum: [PENDING, APPROVED, REJECTED]
        visibility:
          type: string
          enum: [PUBLIC, PRIVATE]
        payload:
          type: object
          description: Activity-specific data
        attachments:
          type: array
          items:
            type: string
        reviewer_id:
          type: string
          nullable: true
        review_note:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - user_id
        - activity_code
        - status
        - payload
...
    Badge:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        description:
          type: string
        icon_url:
          type: string
        criteria:
          type: object
        created_at:
          type: string
          format: date-time
      required:
        - id
        - code
        - name
    EarnedBadge:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
        badge_id:
          type: string
        earned_at:
          type: string
          format: date-time
        badge:
          $ref: '#/components/schemas/Badge'
      required:
        - id
        - user_id
        - badge_id
        - earned_at
...
paths:
...
  /dashboard:
    get:
      summary: Get current user's dashboard data
      tags: [User]
      responses:
        '200':
          description: User dashboard information
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  submissions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Submission'
                  totalPoints:
                    type: integer
                  earnedBadges:
                    type: array
                    items:
                      $ref: '#/components/schemas/EarnedBadge'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /emails/welcome:
    post:
      summary: Send welcome email
      tags: [Email]
      security: []
      description: Internal email service endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                to:
                  type: string
                  format: email
                name:
                  type: string
              required:
                - to
                - name
      responses:
        '200':
          description: Email sent successfully
  /emails/approval:
    post:
      summary: Send submission approval email
      tags: [Email]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                to:
                  type: string
                  format: email
                submissionType:
                  type: string
                pointsAwarded:
                  type: integer
              required:
                - to
                - submissionType
                - pointsAwarded
      responses:
        '200':
          description: Email sent successfully
  /emails/rejection:
    post:
      summary: Send submission rejection email
      tags: [Email]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                to:
                  type: string
                  format: email
                submissionType:
                  type: string
                reason:
                  type: string
              required:
                - to
                - submissionType
      responses:
        '200':
          description: Email sent successfully
...
  /admin/users:
    get:
      summary: Get users list for management
      tags: [Admin]
      security:
        - ClerkJWT: []
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [participant, reviewer, admin, superadmin]
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Paginated users list
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/User'
                        - type: object
                          properties:
                            _count:
                              type: object
                              properties:
                                submissions:
                                  type: integer
                            pointsTotal:
                              type: integer
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      pages:
                        type: integer
...